+++ CHANGELOG.md
@@ -31,6 +31,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - [#2079](https://github.com/clj-kondo/clj-kondo/issues/2079): analysis for `data_readers.clj`
 - [#2067](https://github.com/clj-kondo/clj-kondo/issues/2067): support `:ns-groups` to be used with `:analyze-call` and `:macroexpand` hooks
 - [#1918](https://github.com/clj-kondo/clj-kondo/issues/1918): ignore keyword bindings with namespaced in `:keyword-binding` linter
+- [#2073](https://github.com/clj-kondo/clj-kondo/issues/2073): `:lint-as clj-kondo.lint-as/def-catch-all` should ignore unresolved namespaces
 
 ## 2023.04.14